Move Sandbox ExampleLayer to separate file
Cleaning up SandboxApp #170
